Example check-uk-visa flow in smartdown
This is a hypothetical sketch of how it might work.

The main file `check-uk-visa.txt` defines the overall metadata of the
flow and the start node. It also has global definitions of predicates.

Each node is defined in a separate file whose name matches the node
name.

Each fle has three parts:

* front-matter defines metadata
* body defines the title and body of the question, plus the question
  type and options (acceptable responses) to the question
* for question nodes a 'next nodes' section defines the logic for
  transitioning to the next node

Outcomes are all in sub-folder

Test scenarios are defined in the `scenarios/` subfolder and follow a
different syntax:

node1: response1
node2: response2
outcome

This allows checking the expected path through the graph for a
particular response sequence.