Ensure markdown_blocks always produces array
this trick of using repeat(1,1) matches one repetition of a clause but
causes the match data produced to always be an array containing one
element. This avoids the wierd edge-case that when there is a single
markdown block you would get the markdown block alone, but when there
are multiple you would get an array. Instead you always get an array.