Bugfix: interpolation needs to work with parslet slices
the parser actually produces parslet 'slices' rather than strings.
A parslet slice behaves like a string but doesn't have all of
the methods of string.

So, interpolate needs to #to_s the thing it receives to be able to
handle slices.

This test is to document that fact and catch any regressions.