Use question name for input variable name
The next-node transition rules have until now been independent from
question definitions. However, now that questions can be named
explicitly then we should use the question name when capturing the input
value.

It was an oversight to have missed this when I implemented the choice
question tag with an explicit question name.