Re-use parser/transform during interpolation
Can't reuse node interpreter, as it resolved a node, whereas an individual predicate depends
on a parent rule to me matched, eg a next node set.

The parse/transform repitition could more refactored away quite easily though, and means
the exact same models/logic can be reused for interpolation without any duplication. So
functions, et al, can be used.

This does require making NamedPredicate non-boolean, as question identifier defines as
optionaly including a ? at the end, which is reused by Named Predicate, meaning _any_
identifier like value ([a-zA-Z_0-9-]) gets parsed as a named predicate.

It named predicate is boolean, as it was before, then theres no way to get at the value
of the state by key, as it will always be a Named predicate.

For now, switch named predicate back to returning the value (which mostly gets treated
as boolean), but the proper fix is re-writing parsing to require named predicates to
have a ? at the end. That needs further thought though.

A better solution would be defining a variable type, which evaluates against state,
however that will impact on other definitions, eg, next node rules point at a node
and node name is based on identifier, which isn't a variable