Refactor NamedPredicate to require a trailing ?
Previously it was defined as an identifier, with an optional trailing '?',
which made it inditinguishable from an identifier, so any identifier would
be parsed as a NamedPredicate and return a boolean, not the value from state

This change means the identifier is not parsed at all, and is rescued by
```Parslet::ParseFailed```
so we can attempt to grab the value from state instead.

Eg,
- "%{hello}" will not parse, so we'll get the value from state instead
- "%{function()}" will parse as a FunctionPredicate, and get evaluated

This duplicates some of the evaluatable-or-state var from Functions[1],
where we could do with a Variable type object, that can be evaluated
and have a parsing rule to create them, so we can just assume everything
is evaluatable

This could break in some nasty ways on badly formed input, but that is
the existing behaviour, eg, a bad state var. So, not a regression

[1] https://github.com/alphagov/smartdown/commit/ff83ecc41d95eb84f1fcbdf54a1d5b39695d01d1#diff-cefc093efd4d3702a287586aae7f0fc0R15