Introduce data modules to Node Transformer
Node interpreter now takes an optional data module,
this is similar to existing plugins, except that
they will be available at Flow build time, opposed
to render time.

Monkey patching parslet was the least painful way
to do this, otherwise we would need to introduce a
notion of partially created Elements and have a
post-transformation step.