Don't insert extra line-breaks, breaking govspeak
The handling of new/blank lines changed in https://github.com/alphagov/smartdown/pull/126

This introduced a bug where extra new lines were being inserted. Both the original
new lines captured in MarkdownBlankLine and MarkdownLine and from the new lines
used when joining Markdown Elements.

Normally additional new lines are collapsed in markdown/govspeak, but some
situations, like tables, the addition new lines will break the markdown.

Now that we capture the new line and the markdown (in Elements) we don't
need to join elements with new lines. Otheriwse an element structure like:
```
	[string, new line, string]
```
which will be joined as
```
	string + new line + new line + new line string
```
The specs for API:Node needed to be updated, as they weren't using
the new Elements object and need to mimic the actual structure of a node

I think this shows that the current Element/Elements new line model
isn't quite right, and not particuarly intuitive and probably needs
revisting at some point, but for new this prevents the bug causing
broken markdown to be generated